requirements.txt
Django==3.2.5
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
numpy==1.22.3
PyJWT==2.3.0